Allow function body ASTs to be GC'd
This moves the function body AST into the RuntimeGeneratedFunction
itself, and caches the function object itself, but only holds it in the
cache via a WeakRef, thus ensuring that it can be GC'd as normal. (A
slight wart with this approach is that the id is still present in the
cache after GC has run.)

As a side effect, this makes accessing the body in normal non-generated
functions a lot neater.