Fix bug with GC and function lifetime in the cache
Functions with bodies which share identical syntax collide in the cache.
The cache needs to be carefully managed to ensure this isn't a problem.

In particular, if a cache entry already exists for a given `id`, we need
to use that as the body rather than the body we have. But only if it
wasn't previously collected.