Ensure globals are looked up in the user's module
Create method of `generated_callfunc` in the user's module so that any
global symbols within the body will be looked up in the user's module
scope.

This is straightforward but clunky.  A neater solution should be to
explicitly expand in the user's module and return a CodeInfo from
`generated_callfunc`, but it seems we'd need `jl_expand_and_resolve`
which doesn't exist until Julia 1.3 or so. See:

* https://github.com/JuliaLang/julia/pull/32902
* https://github.com/NHDaly/StagedFunctions.jl/blob/master/src/StagedFunctions.jl#L30